[PATCH] use wait_task_inactive() in kthread_bind()
From: Rusty Russell <rusty@rustcorp.com.au>

Make it clear that the reason we do wait_task_inactive is because
kthread_bind frobs with k->thread_info->cpu, which is only legal because
the task is definitely not running.

We can't use the normal migration thread code here, because it doesn't let
us bind to cpus which are offline yet, and also because we use this to
start the migration threads themselves.

BKrev: 4053d0d1OFfcPae8Q7T-OGdz0Uu75Q