[PATCH] Work around an AMD768MPX erratum
From: Andi Kleen <ak@suse.de>

This patch has been in the SuSE 2.4 kernel forever, but for some reason
never made it mainline.

It works around the infamous "only works stable when a mouse is plugged in"
problem some AMD 768MPX Dual Athlon chipsets have.  The problem happens
because the chipset can hang when PCI prefetch strides from a RAM page into
the VGA text buffer.  When a PS2 mouse is plugged in the BIOS reserves a
page before the VGA text buffer, which stops the prefetch early.

This patch always reserves this page when the chipset could be AMD768MPX.
This can be only done early in bootmem setup.  Because it's difficult to
scan the PCI bus that early it just always reserves this page when the CPU
is an Athlon.  Normally it should not make a difference because the BIOS
will have reserved that page anyways when a PS/2 mouse is plugged in.

BKrev: 4053d0efQlHCzGZaUSHPRRG4WKiOdQ