Merge bk://linux-scsi.bkbits.net/scsi-for-linus-2.6 into ppc970.osdl.org:/home/torvalds/v2.5/linux
2004/03/14 00:40:13-05:00 James.Bottomley
[PATCH] Add Domain Validation to 53c700 driver

This simply throws out the 53c700 driver's optimistic setting of the
best possible transport parameters and replaces it with DV
determination.  It also adds a missing report_bus_reset() callback and
finally does a delayed DV on device errors.

2004/03/13 13:48:36-05:00 James.Bottomley
[PATCH] Fix removable USB drive oops

The actual problem reported was because there wasn't a corresponding
check on transport_classdev.class in the unregister.

However, on closer inspection I also turned up a nasty thinko in the
reference counting.  For reasons best known to the class code authors,
class devices have to obtain their own references to the devices they're
attached to which they release again in their .release routines, so you
have to remember to do a get_device() in the correct place after the
class_device_add().  I put comments in the code so that, hopefully, we
can avoid the problem in future.

2004/03/13 10:35:20-05:00 James.Bottomley
[PATCH] Add Domain Validation to the SPI transport class

Domain Validation is a fairly essential element to the SCSI Parallel
Interface (although if you look very few drivers actually do it).  The
premise is that the Parallel Bus, being a transmission line, might not
be correctly tuned to the transfers you want do perform.  DV probes the
parameters of the transport until it finds a setting that works (for the
interested, see http://www.t10.org/ftp/t10/drafts/sdv/sdv-r08b.pdf)

The current code employs rather simplistic DV heuristics, although those
can be improved over time.  The change in scsi_scan.c is so that DV may
be done easily from the slave_configure routine, which is the most
natural place to begin.

2004/03/12 17:00:50-05:00 James.Bottomley
[PATCH] update the 53c700 use of transport attributes

This patch just brings it up to date with the previous transport
attribute patch, moving it to the model where it sets the min/max of the
attribute if asked for something outside its range.  It also only makes
period and offset visible (it doesn't care about any of the others).

2004/03/12 16:59:27-05:00 James.Bottomley
[PATCH] more SPI transport attribute updates

This does three things

- Fix the signedness of the bit attributes (otherwise they show up as -1
when on, not 1)
- Make the period adjust to the closest value rather than ignoring
values it doesn't understand.
- Add a visibility field to attributes, so drivers can get rid of
attributes they're never going to care about

2004/03/12 16:52:52-05:00 James.Bottomley
[PATCH] add device quiescing to the SCSI API

This patch adds the ability to quiesce a SCSI device.  The idea is that
user issued commands (including filesystem ones) would get blocked,
while mid-layer and device issued ones would be allowed to proceed.
This is for things like Domain Validation which like to operate on an
otherwise quiet device.

There is one big change: to get all of this to happen correctly,
scsi_do_req() has to queue on the *head* of the request queue, not the
tail as it was doing previously.  The reason is that deferred requests
block the queue, so anything needing executing after a deferred request
has to go in front of it.  I don't think there are any untoward
consequences of this.

2004/03/12 16:29:49-05:00 markh
[PATCH] add adapter support to aacraid driver (update)

Mark S. said that there was another adapter added, and that they changed
the names of some boards.  Here is the updated version.

2004/03/12 16:28:27-05:00 willy
[PATCH] sym2 2.1.18i

 - Correct a typo "mvram" -> "nvram".
 - Re-do the PQS/PDS support which I'd #if 0 out.  Should even work on
   multiple-domain boxes now ;-)
 - Move all the nvram definitions to sym_nvram.h (from Gerard's 2.1.19-pre3)
 - hcb_p -> struct sym_hcb *
 - sdev_p -> struct sym_device *
 - Delete a lot of unused macros from sym_misc.h
 - Move READ_BARRIER and WRITE_BARRIER definitions to sym_glue.h
 - SYM_CONF_NVRAM_WRITE_SUPPORT (from Gerard's 2.1.19-pre3).  Not enabled
   yet.
 - Fix some -W warnings (some courtesy of Adrian Bunk).

2004/03/12 16:27:04-05:00 markh
[PATCH] aacraid driver patch

I submitted a patch last month for the aacraid driver's reset handler.
I left out setting function pointers in the adapter_ops structure for
the adapter_check_health element.

2004/03/12 16:26:21-05:00 jejb
MPT Fusion driver 3.01.01 update

From: 	Moore, Eric Dean <Emoore@lsil.com>

This is an update for the MPT Fusion drivers 2.6 kernel.
Version 3.01.01.

This is a fix for poor performance in RAID Volumes.
The dvStatus was being cleared for hidden physical disks
when mptscsih_slave_destroy is called.

Also, I have fixed the warning comming from mptscsih_reset_timeouts.

BKrev: 405491e6Q81pUEeTAjxHk41zNY_brw