[PATCH] Set module owner in megaraid driver
This bug was discussed on linux-scsi a few weeks back, but it appears to
have slipped through the cracks. This fix was originally proposed by
Christoph Hellwig, but I am reposting it, since I have the hardware :-)

Anyway, the problem that this fixes is that megaraid doesn't set the
moduler owner for the host_template, this means that the module can be
removed, even when it is in use.

BKrev: 40568d62YFvnRmE7HPRYJymszvwYag