Merge http://lia64.bkbits.net/to-linus-2.5 into ppc970.osdl.org:/home/torvalds/v2.5/linux
2004/03/16 07:07:38-08:00 benh
[PATCH] g5: Fix iommu vs. pci_device_to_OF_node

The g5 iommu code would fill the "iommu_table" member of whatever
device node was pointed to by pcidev->sysdata during boot. However,
the ppc64 kernel fills that with a pointer to the PHB node which is
later replaced "lazily" with a pointer to the real node when calling
pci_device_to_OF_node(). In this case, we were thus "losign" the
iommu_table pointer. Typical symptom: loss of the SATA when looking
at it's /proc entry.

This fixes it by forcing the update to the final sysdata pointer
when filling up the iommu_table pointers. The "lazy" thing is useless
on pmac anyway.

BKrev: 40571865Y3p4UbigMSfHGZo_CxY5Jg