[PATCH] ppc64: fix for massive OF properties
From: Anton Blanchard <anton@samba.org>

We have some versions of firmware out there that have huge OF properties.
So huge that we end up overwriting our initrd.

Place a 1MB limit and warn bitterly if its over this.  Also fix a use of
package-to-path where the variable was 64bytes but we would pass in a
length of 255.

BKrev: 40571d61zcU5hflqJz_rsOztvidG0g