[PATCH] x86 vsyscall alignment fix
From: Andy Whitcroft <andyw@uk.ibm.com>

The vsyscall implementation for ia32 provides two different vsyscall pages;
one to use int80 and the other to use sysenter.  Each includes a common
signal trailer.  The kernel requires the alignment of routines in this
trailer be consistent in both copies.  However this is not enforced at
compile time.  Failure to maintain this alignment typically leads to an
obscure SIGSEGV in init during boot.  This critical alignment requirement
is not well documented.

Below is a patch to better document the alignment requirements and to
enforce the requirement.

BKrev: 40586a0aeG8uzpYf14fABpV7zkb2lg