[PATCH] config: choice fix
From: Roman Zippel <zippel@linux-m68k.org>

When a boolean choice value has a dependency of 'm' it can be shortly
treated as a tristate symbol.  This fixes this and also add a small
optimization to precompute the value of the module symbol instead of
checking it all the time.

BKrev: 405b0b1f84BfFlRaLbA88xVW9sf8Sg