[PATCH] ppc32: fix build with CONFIG_MODVERSIONS
From: Paul Mackerras <paulus@samba.org>
      Olaf Hering <olh@suse.de>

This adds __kcrctab to the discard list for the link script for the boot
wrapper, and is needed if you compile with CONFIG_MODVERSIONS.

With this patch applied, we need to update makefile dependencies.  Otherwise
the bootfiles will remain untouched because they do not depend on the changed
linker script.

BKrev: 40605617OuFL5uLkZFl4c2Y8bD_Xww