[PATCH] adjuct cpu_khz in response to cpufreq changes
From: john stultz <johnstul@us.ibm.com>

This patch ensures that cpu_khz is adjusted on cpufreq notifications even
when the tsc timesource is not in use.  It fixes the mostly cosmetic issue
when using the ACPI PM timesource of /proc/cpuinfo not being properly
updated when cpu frequency was lowered.

BKrev: 4065ebc3MF9I-4MHcuU-JbpOeGHmiQ