Re-instate __ARCH_SI_BAND_T, to allow architecture overrides.
We default to "long" (which is what POSIX says), but since
a number of architectures have used "int" for historical
reasons, we need to allow overrides.

At least sparc64 needs this. Possibly others, but so far
architecture maintainers haven't spoken up. ppc64 and x86-64
are known to be ok with the default "long".

BKrev: 406f5a3dIqz7HI94_2QUIXH_mvE8mA