[PATCH] remove down_tty_sem()
Remove the down_tty_sem() and up_tty_sem() and replace them with open-coded
up() and down().  This is an equivalent transformation.

I assume these functions were created to open the possibility of per-tty
semaphores at some time in the future.  But the code which is protected by
this lock deals with two tty's at the same time, and the next patch will need
to release the lock after the tty has been destroyed.

BKrev: 407af0baXERhYmrKpyrX1w70R_cQSg