[PATCH] ppc64: Move EPOW log buffer to BSS
From: Olof Johansson <olof@austin.ibm.com>

RTAS on IBM pSeries runs in real mode, so all pointers being passed in to
it need to be in low memory.  There's two places in the RAS code that
passes in pointers to items on the stack, which might end up being above
the limit.

Below patch resolves this by creating a buffer in BSS + a lock for
serialization.  There's no reason to worry about contention on the lock,
since rtas_call() also serializes on a single spinlock and this is an
infrequent code path in the first place.

BKrev: 407af176Oy8d9K19wpNleRl-_CfI4w