[PATCH] selinux: fix struct type
From: Stephen Smalley <sds@epoch.ncsc.mil>

This patch fixes the type of the ssec pointer in the sk_free_security
function.  This has no current impact as the magic element is the top of each
structure.  Thanks to Chad Hanson of TCS for discovering the bug and
submitting the patch.

BKrev: 407af6f1JYDawPHRcWu7wTuRLsEY8A