[PATCH] revert the slabification of i386 pgd's and pmd's
This code is playing with page->lru from pages which came from slab.  But to
remove page->list we need to convert slab over to using page->lru.  So we
cannot allow the i386 pagetable code to go scribbling on the ->lru field of
active slab pages.

This optimisation was pretty thin, and it is more important to shrink the
pageframe (on all architectures).

BKrev: 407af82e3AJ1T-aaK6so8zDD1W6BLA